cloud_function/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.0.0
google-api-python-client>=2.0.0
google-generativeai>=0.8.3
pypdf
pymupdf
google-auth>=2.0.0
google-cloud-tasks>=2.0.0
